Halfway working implementation of streamers
Decryption seems to work, but I have not implemented
keyfile support yet. Should be straight forward.

Have not really tested encryption yet.

Overall impression: I do like this format better,
but hard to tell if it is actually better or not.
It was pretty quick to put together a working
decryption scheme, so that is good. Also feels
like more managable pieces, at the cost of more
total pieces.